agent-service/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "RAG-based Elevator JD Analyzer"
authors = ["Vipshal wg <vipshalgw@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
langgraph = "^0.6.7"
fastapi = "^0.117.1"
uvicorn = "^0.37.0"
google-genai = "^1.39.1"
python-multipart = "^0.0.20"
pypdf2 = "^3.0.1"
python-docx = "^1.2.0"
spacy = "^3.8.7"
nltk = "^3.9.1"
scikit-learn = "^1.7.2"
pydantic = "^2.11.9"
beanie = "^2.0.0"
motor = "^3.7.1"
python-decouple = "^3.8"
python-dotenv = "^1.1.1"
cryptography = "^46.0.1"
python-jose = "^3.5.0"

[tool.poetry.scripts]
server = "backend.server:start_server"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
